move logger, pseudorandom, wrapper_interfaces to framework/utils/
