+++ contracts/integrations/test/framework/assertions/function_assertion.ts
@@ -2,7 +2,7 @@ import { BaseContract, ContractFunctionObj, ContractTxFunctionObj } from '@0x/ba
 import { TransactionReceiptWithDecodedLogs, TxData } from 'ethereum-types';
 import * as _ from 'lodash';
 
-import { logger } from '../logger';
+import { logger } from '../utils/logger';
 
 // tslint:disable:max-classes-per-file
 export type GenericContractFunction<T> = (...args: any[]) => ContractFunctionObj<T>;
