+++ contracts/integrations/test/framework/tests/deployment_manager_test.ts
@@ -2,7 +2,7 @@ import { constants as stakingConstants } from '@0x/contracts-staking';
 import { blockchainTests, expect } from '@0x/contracts-test-utils';
 
 import { DeploymentManager } from '../deployment_manager';
-import { Authorizable, Ownable } from '../wrapper_interfaces';
+import { Authorizable, Ownable } from '../utils/wrapper_interfaces';
 
 blockchainTests('Deployment Manager', env => {
     let owner: string;
